requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
fonttools==4.38.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.29
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
importlib-metadata==5.1.0
ipykernel==6.19.2
ipython==8.7.0
ipywidgets==8.0.3
jcopdl==2.3.0
jcopml==1.1.7
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyterlab-widgets==3.0.4
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mlxtend==0.21.0
multidict==6.0.4
multimethod==1.9
nest-asyncio==1.5.6
networkx==2.8.8
numpy==1.23.5
openai==0.26.0
packaging==22.0
pandas==1.5.2
pandas-profiling==3.5.0
parso==0.8.3
patsy==0.5.3
phik==0.12.3
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.11
platformdirs==2.6.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyaml==21.10.1
pyarrow==10.0.1
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
rich==12.6.0
scikit-learn==1.2.0
scikit-optimize==0.9.0
scipy==1.9.3
seaborn==0.12.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
statsmodels==0.13.5
stqdm==0.0.5
streamlit==1.29.0
streamlit-chat==0.0.2.1
streamlit-option-menu==0.3.2
streamlit-pandas-profiling==0.1.3
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
typeguard==2.13.3
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
visions==0.7.5
watchdog==2.2.0
wcwidth==0.2.5
widgetsnbextension==4.0.4
wincertstore==0.2
yarg==0.1.9
yarl==1.8.2
zipp==3.11.0
